error handling
try-catch

+++ error_handling/employee_model.php
@@ -7,20 +7,41 @@ class employee_model
 
 	public function delete($data)
 	{
+		$count = 0;
 		$hostname = "localhost";
 		$username = "root";
 		$password = "";
 		$database ="php_basics";
-		$dbcon = mysql_connect($hostname, $username, $password) or die ("ko ket noi den mysql");
-		$selected = mysql_select_db($database,$dbcon) or die("ko the lay php_basics");
-		//check connect: error: no connect
-		$result = mysql_query("select count(id) as id from employee where id = '".$data['id']."'");
-		$count = mysql_fetch_assoc($result);
-		echo $count['id'];
-		//check id exit : if ($count['id'] == 0 ) echo "id no exit";
-		$result = mysql_query("DELETE FROM employee WHERE id = '".$data['id']."'");
-		$count = mysql_affected_rows();
-		mysql_close($dbcon);
+		try
+		{
+			$dbcon = mysql_connect($hostname, $username, $password) or die ("ko ket noi den mysql");
+			if($dbcon == false)
+			{	
+				throw new Exception('no connect');
+			}
+			$selected = mysql_select_db($database,$dbcon) or die("ko the lay php_basics");
+			if($selected == false)
+			{
+				throw new Exception('no name data');
+			}
+			$result = mysql_query("select count(id) as id from employee where id = '".$data['id']."'");
+			$test = mysql_fetch_assoc($result);
+			if($test['id'] != 1)
+			{
+				throw new Exception('id no exit');
+			}
+			$result = mysql_query("DELETE FROM employee WHERE id = '".$data['id']."'");
+			if(!isset($result))
+			{
+				throw new Exception('delete no access');
+			}
+			$count = mysql_affected_rows();
+			mysql_close($dbcon);
+		}
+		catch(Exception $e)
+		{
+			echo $e->getmessage();
+		}
 		return $count;
 		
 	}// kiem tra id ton tai: check_id_exit()
@@ -29,50 +50,121 @@ class employee_model
 
 	public function insert($data)
 	{
+		$count = 0;
 		$hostname = "localhost";
 		$username = "root";
 		$password = "";
 		$database ="php_basics";
-		$dbcon = mysql_connect($hostname, $username, $password) or die ("ko ket noi den mysql");
-		$selected = mysql_select_db($database,$dbcon) or die("ko the lay php_basics");
-		$sql = "INSERT INTO employee (name , title, created, modified) VALUES ('".$data['name']."','".$data['title']."' , '".$data['created']."', '".$data['modified']."')";
-		$result = mysql_query($sql);
-		echo mysql_error();
-		$result = mysql_insert_id();
-		mysql_close($dbcon);
-		return $result;
+		try
+		{
+			$dbcon = mysql_connect($hostname, $username, $password) or die ("ko ket noi den mysql");
+			if($dbcon == false)
+			{	
+				throw new Exception('no connect');
+			}
+			$selected = mysql_select_db($database,$dbcon) or die("ko the lay php_basics");
+			if($selected == false)
+			{
+				throw new Exception('no name data');
+			}
+			$sql = "INSERT INTO employee (name , title, created, modified) VALUES ('".$data['name']."','".$data['title']."' , '".$data['created']."', '".$data['modified']."')";
+			$result = mysql_query($sql);
+			if(!isset($result))
+			{
+				throw new Exception('insert no access');
+			}
+			$count = mysql_insert_id();
+			mysql_close($dbcon);
+		}
+		catch(Exception $e)
+		{
+			echo $e->getmessage();
+		}
+		return $count;
 	}
 
 /////////////UPDATE/////////////////
 
 	public function update($data)
 	{
+		$count = 0;
 		$hostname = "localhost";
 		$username = "root";
 		$password = "";
 		$database = "php_basics";
-		$dbcon = mysql_connect($hostname, $username, $password) or die (" no connect");
-		$selected = mysql_select_db($database, $dbcon) or die ("no database");
-		$sql = "update employee set name = '".$data['name']."', title = '".$data['title']."' , modified = '".$data['modified']."' where id = '".$data['id']."'";
-		$result = mysql_query($sql);
-		$count = mysql_affected_rows();
-		mysql_close($dbcon);
+		try
+		{
+			$dbcon = mysql_connect($hostname, $username, $password) or die ("ko ket noi den mysql");
+			if($dbcon == false)
+			{	
+				throw new Exception('no connect');
+			}
+			$selected = mysql_select_db($database,$dbcon) or die("ko the lay php_basics");
+			if($selected == false)
+			{
+				throw new Exception('no name data');
+			}
+			$result = mysql_query("select count(id) as id from employee where id = '".$data['id']."'");
+			$test = mysql_fetch_assoc($result);
+			if($test['id'] != 1)
+			{
+				throw new Exception('id no exit');
+			}
+			$sql = "update employee set name = '".$data['name']."', title = '".$data['title']."' , modified = '".$data['modified']."' where id = '".$data['id']."'";
+			$result = mysql_query($sql);
+			if(!isset($result))
+			{
+				throw new Exception('insert no access');
+			}
+			$count = mysql_affected_rows();
+			mysql_close($dbcon);
+		}
+		catch(Exception $e)
+		{
+			echo $e->getmessage();
+		}
 		return $count;
 	}
 
 /////////////SELECT_BY_ID////////////
 
-	public function select_by_id($id)
+	public function select_by_id($data)
 	{
+		$row = 0;
 		$hostname = "localhost";
 		$username = "root";
 		$password = "";
 		$database ="php_basics";
-		$dbcon = mysql_connect($hostname, $username, $password) or die ("ko ket noi den mysql");
-		$selected = mysql_select_db($database,$dbcon) or die("ko the lay php_basics");
-		$result = mysql_query("SELECT * FROM employee WHERE id=$id");
-		$row = mysql_fetch_array($result, MYSQL_ASSOC);
-		mysql_close($dbcon);
+		try
+		{
+			$dbcon = mysql_connect($hostname, $username, $password) or die ("ko ket noi den mysql");
+			if($dbcon == false)
+			{	
+				throw new Exception('no connect');
+			}
+			$selected = mysql_select_db($database,$dbcon) or die("ko the lay php_basics");
+			if($selected == false)
+			{
+				throw new Exception('no name data');
+			}
+			$result = mysql_query("select count(id) as id from employee where id = '".$data['id']."'");
+			$test = mysql_fetch_assoc($result);
+			if($test['id'] != 1)
+			{
+				throw new Exception('id no exit');
+			}
+			$result = mysql_query("SELECT * FROM employee WHERE id='".$data['id']."'");
+			if(!isset($result))
+			{
+				throw new Exception('select by id no access');
+			}
+			$row = mysql_fetch_array($result, MYSQL_ASSOC);
+			mysql_close($dbcon);
+		}
+		catch(Exception $e)
+		{
+			echo $e->getmessage();
+		}
 		return $row;
 	}
 ///////////VALIDATION///////////////
@@ -160,5 +252,33 @@ class employee_model
 		}
 		return 0;
 	}
+	public function data()
+	{
+		$hostname = "localhost";
+		$username = "root";
+		$password = "";
+		$database = "php_basics";
+		try
+		{
+			$dbcon = mysql_connect($hostname, $username, $password);
+			if($dbcon == false)
+			{
+				throw new Exception('no connect');
+			}
+			$selected = mysql_select_db($database, $dbcon);
+			if($selected ==false)
+			{
+				throw new Exception('no name data');
+			}
+			if($selected == true)
+			{ echo 'access';}
+			mysql_close($dbcon);
+		}
+		catch(Exception $e)
+		{
+			echo $e->getmessage();
+		}
+	}
+	
 
 }
